refactor: consolidate EXIF and ODM processing tests into comprehensive suite
- Removed outdated EXIF extraction tests and integration tests to streamline the testing framework.
- Introduced a new consolidated test suite for ODM processing that covers EXIF extraction, RTK detection, and orthomosaic generation.
- Enhanced test coverage by ensuring all processing stages are validated in a single comprehensive test, improving maintainability and reducing redundancy.
- Updated test fixtures to support the new structure and ensure reliable dataset handling during tests.
